refactor: move textual app to app.py
also adds vscode directory to git ignore
